requirements.txt
anthropic>=0.25.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
pyyaml>=6.0
pandas>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
responses>=0.23.0
PyPDF2>=3.0.0